Changes with Sale APIs
